Add Book entity and service

+++ bookstore/src/main/java/com/example/bookstore/entity/Book.java
@@ -1,14 +1,13 @@
 package com.example.bookstore.entity;
 
+import java.math.BigDecimal;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Column;
 import jakarta.persistence.Table;
 import lombok.Data;
-
-import java.math.BigDecimal;
 
 @Entity
 @Table(name = "books")
@@ -38,7 +37,7 @@ public class Book {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -89,6 +88,7 @@ public class Book {
     public void setCoverImage(String coverImage) {
         this.coverImage = coverImage;
     }
+
     public Book() {
 
     }
